Flatten the structure of constraints
Instead of having a `Constraint` which consists of some `YValueConstraint`,
now there's just plain `Constraint` that knows for which X and Y values
it has to put constraint.

It's needed because new types of constraints that are going to be added,
will need access to the X values, and they will use some non-trivial
assertion logic (not just "for every X value, check some y-value assertion").

Part of #5.