Keep all X values for given column in one constraint
It's done with >1 sampling in mind. When there's more than one
sample per column, it's semantically more correct to say:
"for all X values that will be generated for this column by sampling
it, apply this constraint" than "for all X values (...) create
separate constraints". Why: because some constraints that are
going to be added, need to have access to all X values for a given
column.

For example, if one wants to check that all Y values for all X
values are between a certain range, but also that the given
constraint cannot be shrunk to a smaller range. It would be impossible
to check it having constraints with just one X value, because it's
legal for such constraint to have some values that are in some smaller
range. It's just there have to be some other X values that fill
the areas near range's boundaries.

Part of #5.